Draw shape of player
Although its shape is a rhombus, the player hitbox is a rectangle as
seen on the surface drawn before drawing the polygon itself. That being
said, this is done to save computational calculations with non rectang
ular shapes. It could be done with any shape using a mask_from_surface
method from pygame in order to remove some pixels from the collision
area as well as sprite.spritecollide to detect collisions among sprites
based on their masks.